+++ crates/aranya-capi-codegen/src/syntax/node.rs
@@ -259,9 +259,43 @@ impl ToTokens for Alias {
     fn to_tokens(&self, tokens: &mut TokenStream) {
         // NB: We do not emit `capi::` attributes.
 
+        // Manually expand opaque so we don't need expansion in cbindgen.
+        if let Some(Opaque { size, align, .. }) = &self.opaque {
+            // Generate fake struct for cbindgen
+            tokens.append_all({
+                let vis = &self.vis;
+
+                let doc = &self.doc;
+                let attrs: Vec<Attribute> = self
+                    .attrs
+                    .iter()
+                    .filter(|attr| attr.path().is_ident("cfg"))
+                    .cloned()
+                    .collect();
+
+                let name = &self.ident;
+
+                quote! {
+                    #[cfg(cbindgen)]
+                    #doc
+                    #[repr(C, align(#align))]
+                    #(#attrs)*
+                    #vis struct #name {
+                        /// This field only exists for size purposes. It is
+                        /// UNDEFINED BEHAVIOR to read from or write to it.
+                        /// @private
+                        __for_size_only: [u8; #size],
+                    }
+                }
+            });
+
+            // `cfg` for regular generated alias below
+            tokens.append_all(quote! {#[cfg(not(cbindgen))]});
+        }
+
         self.doc.to_tokens(tokens);
         tokens.append_all(self.attrs.outer());
-        self.opaque.to_tokens(tokens);
+        // `opaque` intentionally omitted, see above.
         self.vis.to_tokens(tokens);
         self.type_token.to_tokens(tokens);
         self.ident.to_tokens(tokens);
@@ -348,11 +382,17 @@ impl ToTokens for Struct {
     fn to_tokens(&self, tokens: &mut TokenStream) {
         // NB: We do not emit `capi::` attributes.
 
+        // Opaque is not currently supported on structs. When supported, we will need to add
+        // similar handling as in `Alias`. We can't seem to remove the field, I think because it is
+        // used to pass an `Opaque` between `Struct` and `Alias`. This panic doesn't seem to be
+        // reachable though, so we correctly never generate a struct with an opaque attribute.
+        assert!(self.opaque.is_none(), "opaque struct: {:?}", self.opaque);
+
         self.doc.to_tokens(tokens);
         self.derives.to_tokens(tokens);
         self.repr.to_tokens(tokens);
         tokens.append_all(self.attrs.outer());
-        self.opaque.to_tokens(tokens);
+        // `opaque` intentionally omitted, see above.
         self.vis.to_tokens(tokens);
         self.struct_token.to_tokens(tokens);
         self.ident.to_tokens(tokens);