+++ crates/aranya-policy-vm/tests/vm.rs
@@ -1103,7 +1103,7 @@ fn test_match_second() -> anyhow::Result<()> {
 }
 
 #[test]
-fn test_match_none() -> anyhow::Result<()> {
+fn test_match_default_2() -> anyhow::Result<()> {
     let name = ident!("foo");
     let policy = parse_policy_str(POLICY_MATCH, Version::V2)?;
     let module = Compiler::new(&policy).compile()?;
@@ -1112,8 +1112,11 @@ fn test_match_none() -> anyhow::Result<()> {
     let ctx = dummy_ctx_action(name.clone());
 
     let mut rs = machine.create_run_state(&io, ctx);
-    let result = rs.call_action(name, [Value::Int(0)])?;
-    assert_eq!(result, ExitReason::Panic);
+    let mut published = Vec::new();
+    call_action(&mut rs, &mut published, name, [100])?.success();
+    drop(rs);
+
+    assert_eq!(published, [vm_struct!(Result { x: 100 + 1 })]);
 
     Ok(())
 }
@@ -1137,6 +1140,7 @@ fn test_match_alternation() -> anyhow::Result<()> {
                 5 | 6 | 7 => {
                     publish Result { x: x }
                 }
+                _ => {}
             }
         }
     "#;
@@ -1203,6 +1207,7 @@ fn test_match_return() -> anyhow::Result<()> {
         function bar() int {
             match 0 {
                 0 => { return 42 }
+                _ => { return 0 }
             }
         }
     "#;
@@ -2096,7 +2101,7 @@ fn test_optional_type_validation() -> anyhow::Result<()> {
     let text = r#"
         command TypeValidation {
             fields {
-                maybe_int optional int,
+                maybe_int option[int],
                 name string,
             }
             seal { return todo() }
@@ -2106,7 +2111,7 @@ fn test_optional_type_validation() -> anyhow::Result<()> {
             }
         }
 
-        action type_validation(maybe_int_input optional int, name_input string) {
+        action type_validation(maybe_int_input option[int], name_input string) {
             publish TypeValidation{maybe_int: maybe_int_input, name: name_input}
         }
     "#;
@@ -2487,5 +2492,42 @@ fn test_struct_conversion() -> anyhow::Result<()> {
             y: Value::String(text!("xyz")),
         })
     );
+    Ok(())
+}
+
+#[test]
+fn test_source_lookup() -> anyhow::Result<()> {
+    let text = r#"
+        action foo() {
+            check true
+            // before
+            check false
+            // after
+            check true
+        }
+    "#;
+
+    let name = ident!("foo");
+    let policy = parse_policy_str(text, Version::V2)?;
+    let io = RefCell::new(TestIO::new());
+    let ctx = dummy_ctx_action(name.clone());
+    let module = Compiler::new(&policy).compile()?;
+    let machine = Machine::from_module(module)?;
+    let mut rs = machine.create_run_state(&io, ctx);
+
+    let result = rs.call_action(name, iter::empty::<Value>())?;
+    assert_eq!(result, ExitReason::Check);
+
+    let source = rs.source_location().expect("could not get source location");
+    assert_eq!(
+        source,
+        concat!(
+            "at row 5 col 13:\n",
+            "\tcheck false\n",
+            "            // after\n",
+            "            "
+        )
+    );
+
     Ok(())
 }