+++ crates/aranya-envelope-ffi/CHANGELOG.md
@@ -7,6 +7,18 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.15.1](https://github.com/aranya-project/aranya-core/compare/aranya-envelope-ffi-v0.15.0...aranya-envelope-ffi-v0.15.1) - 2025-11-12
+
+### Other
+
+- updated the following local packages: aranya-policy-vm
+
+## [0.15.0](https://github.com/aranya-project/aranya-core/compare/aranya-envelope-ffi-v0.14.0...aranya-envelope-ffi-v0.15.0) - 2025-11-05
+
+### Other
+
+- use tagged ID type ([#327](https://github.com/aranya-project/aranya-core/pull/327))
+
 ## [0.14.0](https://github.com/aranya-project/aranya-core/compare/aranya-envelope-ffi-v0.13.0...aranya-envelope-ffi-v0.14.0) - 2025-10-16
 
 ### Other